student-Assistant/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
billiard==4.2.1
celery[redis]==5.3.6
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-tailwind==1.0.3
distro==1.9.0
Django==5.1.4
django-browser-reload==1.17.0
django-crispy-forms==2.3
django-recaptcha==4.0.0
django-storages==1.14.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
kombu==5.4.2
openai==1.59.5
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.48
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
replicate==1.0.4
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
django-storages[s3]

gunicorn==20.1.0
psycopg2==2.8.6
sqlparse==0.4.2
whitenoise==6.2.0